+++ Modules/core/Model/CoreUser.php
@@ -336,8 +336,8 @@ class CoreUser extends Model {
     }
 
     public function edit($id, $login, $name, $firstname, $email, $status_id, $date_end_contract, $is_active) {
-        $sql = "UPDATE core_users SET login=?, name=?, firstname=?, email=?, status_id=?, date_end_contract=?, is_active=? WHERE id=?";
-        $this->runRequest($sql, array($login, $name, $firstname, $email, $status_id, $date_end_contract, $is_active, $id));
+        $sql = "UPDATE core_users SET name=?, firstname=?, email=?, status_id=?, date_end_contract=?, is_active=? WHERE id=?";
+        $this->runRequest($sql, array($name, $firstname, $email, $status_id, $date_end_contract, $is_active, $id));
     }
 
     public function isUserId($id) {